backend/pyproject.toml
[project]
name = "echoshaper-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "openai",
  "PyYAML",
  "pytest",
  "python-dotenv",
  "websockets",
  "python-multipart",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

